Correct normal direction in qhull when exporting to trimesh
When transforming a RW QHull into a triangle mesh currently the normal direction of the remaining triangles is not set to the one determined by qhull.

There exists a "rule" how the normal direction should relate to the individual vertices, therefore just setting the computed normal is not good enough, but rather two of the triangles vertices need to be switched if the direction is 180 deg off.

Could probably be done more efficient by doing the checks locally. This seemed safer to me. If this is used for time critical things (?) one could make this optional using a flag.